Commands to a new class.
