+++ Source/InRoomChat.cs
@@ -4,6 +4,7 @@ using GGM.Config;
 using System;
 using System.Collections.Generic;
 using System.Text;
+using GGM;
 using GGM.GUI.Pages;
 using UnityEngine;
 using MonoBehaviour = Photon.MonoBehaviour;
@@ -235,16 +236,6 @@ public class InRoomChat : MonoBehaviour
         }
     }
 
-    public static bool MCRequired()
-    {
-        if (!PhotonNetwork.isMasterClient)
-        {
-            SystemMessageLocal(Error(0));
-        }
-
-        return !PhotonNetwork.isMasterClient;
-    }
-
     private void Awake()
     {
         Chat = this;
@@ -255,79 +246,26 @@ public class InRoomChat : MonoBehaviour
         switch (args[0])
         {
             case "pos":
-                {
-                    string[] msg = { "Your position:\n", "\nX", " - ", $"{GameObjectCache.Find("MainCamera").GetComponent<IN_GAME_MAIN_CAMERA>().main_object.transform.position.x.ToString()}" + "\nY", " - ", $"{GameObjectCache.Find("MainCamera").GetComponent<IN_GAME_MAIN_CAMERA>().main_object.transform.position.y.ToString()}" + "\nZ", " - ", $"{GameObjectCache.Find("MainCamera").GetComponent<IN_GAME_MAIN_CAMERA>().main_object.transform.position.z.ToString()}" };
-
-                    SystemMessageLocal(msg);
-                }
+                Commands.GetPosition();
                 break;
 
             case "ban":
-                {
-                    if (MCRequired()) return;
-
-                    var id = Convert.ToInt32(args[1]);
-
-                    if (id == PhotonNetwork.player.ID)
-                    {
-                        SystemMessageLocal(Error(2, "ban"));
-                    }
-                    else if (!(FengGameManagerMKII.OnPrivateServer || PhotonNetwork.isMasterClient))
-                    {
-                        FengGameManagerMKII.FGM.photonView.RPC("Chat", PhotonTargets.All, "/kick #" + Convert.ToString(id), LoginFengKAI.player.name);
-                    }
-                    else
-                    {
-                        foreach (var player in PhotonNetwork.playerList)
-                        {
-                            if (id == player.ID)
-                            {
-                                if (FengGameManagerMKII.OnPrivateServer)
-                                {
-                                    FengGameManagerMKII.FGM.kickPlayerRC(player, true, "");
-                                }
-                                else if (PhotonNetwork.isMasterClient)
-                                {
-                                    FengGameManagerMKII.FGM.kickPlayerRC(player, true, "");
-                                    SystemMessageGlobal(player, "has been banned.");
-                                }
-                            }
-                        }
-
-                        if (PhotonPlayer.Find(id) == null)
-                        {
-                            SystemMessageLocal(Error(1));
-                        }
-                    }
-                }
+                Commands.Ban(args[1]);
                 break;
 
             case "aso":
                 switch (args[1])
                 {
                     case "damage":
-                        Settings.CustomSizeSetting.Value = true;
-                        Settings.SizeSettings[0].Value = 100f;
-                        Settings.SizeSettings[0].Value = 0f;
-                        Settings.SizeSettings[0].Value = 0f;
-                        Settings.SizeSettings[0].Value = 0f;
-                        Settings.SizeSettings[0].Value = 0f;
-                        Settings.ArmorModeSetting.Value = true;
-                        Settings.ArmorSetting.Value = 1000;
-                        Settings.CustomSizeSetting.Value = true;
-                        Settings.SizeSettings[0].Value = 2.5f;
-                        Settings.SizeSettings[1].Value = 3f;
-                        SystemMessageGlobal("ASO Damage enabled.");
+                        Commands.ASODamage();
                         break;
 
                     case "kdr":
-                        RCSettings.asoPreservekdr = RCSettings.asoPreservekdr == 0 ? 1 : 0;
-                        SystemMessageGlobal("KDRs will " + (RCSettings.asoPreservekdr == 1 ? string.Empty : "not ") + "be preserved from disconnects.");
+                        Commands.ASOKDR();
                         break;
 
                     case "racing":
-                        RCSettings.racingStatic = RCSettings.racingStatic == 0 ? 1 : 0;
-                        SystemMessageLocal("Restart required.");
+                        Commands.ASORacing();
                         break;
 
                     default:
@@ -335,138 +273,65 @@ public class InRoomChat : MonoBehaviour
                         SystemMessageLocal(err);
                         break;
                 }
-
                 break;
 
             case "clean":
             case "clear":
-                {
-                    for (var i = 0; i < 15; i++)
-                    {
-                        SystemMessageLocal(string.Empty);
-                    }
-                }
+
+                Commands.CleanChat();
+
                 break;
 
             case "/clean":
             case "/clear":
-                {
-                    for (var i = 0; i < 15; i++)
-                    {
-                        SystemMessageGlobal(string.Empty);
-                    }
-                }
+
+                Commands.CleanChat(false);
+
                 break;
 
             case "pause":
             case "unpause":
-                {
-                    if (MCRequired()) return;
-
-                    FengGameManagerMKII.FGM.SetPause();
-                }
+                FengGameManagerMKII.FGM.SetPause();
                 break;
 
             case "ignorelist":
-                foreach (var id in FengGameManagerMKII.ignoreList)
-                {
-                    SystemMessageLocal(id.ToString());
-                }
-
+                Commands.IngoreList();
                 break;
 
             case "slots":
-                {
-                    if (MCRequired()) return;
-
-                    var slots = Convert.ToInt32(args[1]);
-                    PhotonNetwork.room.maxPlayers = slots;
-                    string[] msg = { "Max players changed to", slots.ToString(), "." };
-                    SystemMessageGlobal(msg);
-                }
+                Commands.SetSlots(Convert.ToInt32(args[1]));
                 break;
 
             case "time":
-                {
-                    if (MCRequired()) return;
-
-                    var time = (FengGameManagerMKII.FGM.time - (int)FengGameManagerMKII.FGM.timeTotalServer - Convert.ToInt32(args[1])) * -1;
-                    FengGameManagerMKII.FGM.addTime(time);
-                    string[] msg = { "Time set to", time.ToString(), "." };
-                    SystemMessageGlobal(msg);
-                }
+                Commands.SetTime(Convert.ToInt32(args[1]));
                 break;
 
             case "tp":
-                {
-                    var player = PhotonPlayer.Find(Convert.ToInt32(args[1]));
-                    var obj = new GameObject();
-                    var obj2 = new GameObject();
-                    var tpPlayers = GameObject.FindGameObjectsWithTag("Player");
-                    for (var i = 0; i < tpPlayers.Length; i++)
-                    {
-                        var obj3 = tpPlayers[i];
-                        if (obj3.GetPhotonView().owner == PhotonPlayer.Find(Convert.ToInt32(args[1])))
-                        {
-                            obj = obj3;
-                        }
-
-                        if (obj3.GetPhotonView().owner == PhotonNetwork.player)
-                        {
-                            obj2 = obj3;
-                        }
-                    }
-
-                    SystemMessageLocal("Teleported to ", player, ".");
-                    obj2.transform.position = obj.transform.position;
-                }
+                Commands.Teleport(Convert.ToInt32(args[1]));
                 break;
 
             case "reconnect":
-                {
-                    FengGameManagerMKII.NeedRejoin = true;
-                    PhotonNetwork.Disconnect();
-                }
+                Commands.Reconnect();
                 break;
 
             case "resetkd":
-                {
-                    PhotonNetwork.player.SetCustomProperties(new Hashtable { { "kills", 0 }, { "deaths", 0 }, { "max_dmg", 0 }, { "total_dmg", 0 } });
-                    SystemMessageLocal("Your stats have been reset.");
-                }
+                Commands.ResetKD();
+                break;
+
+            case "/resetkd":
+                Commands.ResetKD(args[1]);
                 break;
 
             case "resetkdall":
-                {
-                    if (MCRequired()) return;
-
-                    var hash = new Hashtable { { "kills", 0 }, { "deaths", 0 }, { "max_dmg", 0 }, { "total_dmg", 0 } };
-                    foreach (var player in PhotonNetwork.playerList)
-                    {
-                        player.SetCustomProperties(hash);
-                    }
-
-                    SystemMessageGlobal("All stats have been reset.");
-                }
+                Commands.ResetKD(global: true);
                 break;
 
             case "revive":
-                {
-                    if (MCRequired()) return;
-
-                    var player = PhotonPlayer.Find(Convert.ToInt32(args[1]));
-                    FengGameManagerMKII.FGM.photonView.RPC("respawnHeroInNewRound", player);
-                    SystemMessageGlobal(player, "has been revived.");
-                }
+                Commands.Revive(args[1]);
                 break;
 
             case "reviveall":
-                {
-                    if (MCRequired()) return;
-
-                    FengGameManagerMKII.FGM.photonView.RPC("respawnHeroInNewRound", PhotonTargets.All);
-                    SystemMessageGlobal("All players have been revived.");
-                }
+                Commands.Revive(all: true);
                 break;
 
             case "pm":
@@ -502,439 +367,33 @@ public class InRoomChat : MonoBehaviour
 
             case "team":
                 {
-                    if (RCSettings.teamMode != 1)
-                    {
-                        string[] msg = { "Teams ", "are locked or disabled." };
-                        SystemMessageLocal(msg, false);
-                        return;
-                    }
-
-                    var teamValue = 0;
-                    var newTeamName = "Individuals";
-                    switch (args[1])
-                    {
-                        case "0":
-                        case "individual":
-                            break;
-
-                        case "1":
-                        case "cyan":
-                            teamValue = 1;
-                            newTeamName = "Cyan";
-                            break;
-
-                        case "2":
-                        case "magenta":
-                            teamValue = 2;
-                            newTeamName = "Magenta";
-                            break;
-
-                        default:
-                            string[] err = { "Invalid team code/name. Possibles:\n" + "Team Individuals - ", "0", "/", "individuals.\n", "Team Cyan - ", "1", "/", "cyan.\n", "Team Magenta - ", "2", "/", "magenta." };
-                            SystemMessageLocal(err);
-                            return;
-                    }
-
-                    FengGameManagerMKII.FGM.photonView.RPC("setTeamRPC", PhotonNetwork.player, teamValue);
-                    string[] msg2 = { "You have joined ", "Team " + newTeamName, "." };
-                    SystemMessageLocal(msg2);
-                    foreach (var obj in FengGameManagerMKII.FGM.getPlayers())
-                    {
-                        var her = (HERO)obj;
-                        if (her.photonView.isMine)
-                        {
-                            her.markDie();
-                            her.photonView.RPC("netDie2", PhotonTargets.All, -1, "Team Switch");
-                            break;
-                        }
-                    }
+                    Commands.SwitchTeam(args[1]);
                 }
                 break;
 
             case "kick":
-                {
-                    if (MCRequired()) return;
-
-                    var num8 = Convert.ToInt32(args[1]);
-                    if (num8 == PhotonNetwork.player.ID)
-                    {
-                        SystemMessageLocal(Error(2, "kick"));
-                    }
-                    else if (!(FengGameManagerMKII.OnPrivateServer || PhotonNetwork.isMasterClient))
-                    {
-                        FengGameManagerMKII.FGM.photonView.RPC("Chat", PhotonTargets.All, "/kick #" + Convert.ToString(num8), LoginFengKAI.player.name);
-                    }
-                    else
-                    {
-                        foreach (var player in PhotonNetwork.playerList)
-                        {
-                            if (num8 == player.ID)
-                            {
-                                if (FengGameManagerMKII.OnPrivateServer)
-                                {
-                                    FengGameManagerMKII.FGM.kickPlayerRC(player, false, "");
-                                }
-                                else if (PhotonNetwork.isMasterClient)
-                                {
-                                    FengGameManagerMKII.FGM.kickPlayerRC(player, false, "");
-                                    SystemMessageGlobal(player, "has been kicked.");
-                                }
-                            }
-                        }
-
-                        if (PhotonPlayer.Find(num8) == null)
-                        {
-                            SystemMessageLocal(Error(1));
-                        }
-                    }
-                }
+                Commands.Kick(args[1]);
                 return;
 
             case "restart":
-                {
-                    if (MCRequired()) return;
-
-                    FengGameManagerMKII.FGM.restartGame(false);
-                    string[] msg = { "MasterClient ", "has restarted the game." };
-                    SystemMessageLocal(msg, false);
-                }
+                Commands.Restart();
                 return;
 
             case "specmode":
-                if ((int)FengGameManagerMKII.settings[245] == 0)
-                {
-                    FengGameManagerMKII.settings[245] = 1;
-                    FengGameManagerMKII.FGM.EnterSpecMode(true);
-                    string[] msg = { "You have entered ", "Spectator ", "mode." };
-                    SystemMessageLocal(msg);
-                }
-                else
-                {
-                    FengGameManagerMKII.settings[245] = 0;
-                    FengGameManagerMKII.FGM.EnterSpecMode(false);
-                    string[] msg = { "You have exited ", "Spectator ", "mode." };
-                    SystemMessageLocal(msg);
-                }
-
-                return;
-
-            case "fov":
-                {
-                    var fov = Convert.ToInt32(args[1]);
-                    Camera.main.fieldOfView = fov;
-                    string[] msg = { "Field of Vision", "set to", fov.ToString(), "." };
-                    SystemMessageLocal(msg, false);
-                }
+                Commands.SpectatorMode();
                 return;
 
             case "spectate":
-                {
-                    var playerid = Convert.ToInt32(args[1]);
-                    foreach (var player in GameObject.FindGameObjectsWithTag("Player"))
-                    {
-                        if (player.GetPhotonView().owner.ID == playerid)
-                        {
-                            Camera.main.GetComponent<IN_GAME_MAIN_CAMERA>().setMainObject(player);
-                            Camera.main.GetComponent<IN_GAME_MAIN_CAMERA>().setSpectorMode(false);
-                            SystemMessageLocal("You are now spectate", player.GetPhotonView().owner);
-                        }
-                    }
-                }
+                Commands.Spectate(Convert.ToInt32(args[1]));
                 return;
 
             case "rules":
-                {
-                    Rules();
-                }
+                Commands.Rules();
                 break;
 
             default:
                 SystemMessageLocal("Unknown command.");
                 break;
-        }
-    }
-
-    private static void Rules()
-    {
-        if (Settings.LegacyChatSetting)
-        {
-            AddLineRC("Currently activated gamemodes:");
-            if (RCSettings.bombMode > 0)
-            {
-                AddLineRC("Bomb mode is on.");
-            }
-            if (RCSettings.teamMode > 0)
-            {
-                if (RCSettings.teamMode == 1)
-                {
-                    AddLineRC("Team mode is on (no sort).");
-                }
-                else if (RCSettings.teamMode == 2)
-                {
-                    AddLineRC("Team mode is on (sort by size).");
-                }
-                else if (RCSettings.teamMode == 3)
-                {
-                    AddLineRC("Team mode is on (sort by skill).");
-                }
-            }
-            if (RCSettings.pointMode > 0)
-            {
-                AddLineRC("Point mode is on (" + Convert.ToString(RCSettings.pointMode) + ").");
-            }
-            if (RCSettings.disableRock > 0)
-            {
-                AddLineRC("Punk Rock-Throwing is disabled.");
-            }
-            if (RCSettings.spawnMode > 0)
-            {
-                AddLineRC("Custom spawn rate is on (" + RCSettings.nRate.ToString("F2") + "% Normal, " + RCSettings.aRate.ToString("F2") + "% Abnormal, " + RCSettings.jRate.ToString("F2") + "% Jumper, " + RCSettings.cRate.ToString("F2") + "% Crawler, " + RCSettings.pRate.ToString("F2") + "% Punk");
-            }
-            if (RCSettings.explodeMode > 0)
-            {
-                AddLineRC("Titan explode mode is on (" + Convert.ToString(RCSettings.explodeMode) + ").");
-            }
-            if (RCSettings.healthMode > 0)
-            {
-                AddLineRC("Titan health mode is on (" + Convert.ToString(RCSettings.healthLower) + "-" + Convert.ToString(RCSettings.healthUpper) + ").");
-            }
-            if (RCSettings.infectionMode > 0)
-            {
-                AddLineRC("Infection mode is on (" + Convert.ToString(RCSettings.infectionMode) + ").");
-            }
-            if (RCSettings.damageMode > 0)
-            {
-                AddLineRC("Minimum nape damage is on (" + Convert.ToString(RCSettings.damageMode) + ").");
-            }
-            if (RCSettings.moreTitans > 0)
-            {
-                AddLineRC("Custom titan # is on (" + Convert.ToString(RCSettings.moreTitans) + ").");
-            }
-            if (RCSettings.sizeMode > 0)
-            {
-                AddLineRC("Custom titan size is on (" + RCSettings.sizeLower.ToString("F2") + "," + RCSettings.sizeUpper.ToString("F2") + ").");
-            }
-            if (RCSettings.banEren > 0)
-            {
-                AddLineRC("Anti-Eren is on. Using Titan eren will get you kicked.");
-            }
-            if (RCSettings.waveModeOn == 1)
-            {
-                AddLineRC("Custom wave mode is on (" + Convert.ToString(RCSettings.waveModeNum) + ").");
-            }
-            if (RCSettings.friendlyMode > 0)
-            {
-                AddLineRC("Friendly-Fire disabled. PVP is prohibited.");
-            }
-            if (RCSettings.pvpMode > 0)
-            {
-                if (RCSettings.pvpMode == 1)
-                {
-                    AddLineRC("AHSS/Blade PVP is on (team-based).");
-                }
-                else if (RCSettings.pvpMode == 2)
-                {
-                    AddLineRC("AHSS/Blade PVP is on (FFA).");
-                }
-            }
-            if (RCSettings.maxWave > 0)
-            {
-                AddLineRC("Max Wave set to " + RCSettings.maxWave.ToString());
-            }
-            if (RCSettings.horseMode > 0)
-            {
-                AddLineRC("Horses are enabled.");
-            }
-            if (RCSettings.ahssReload > 0)
-            {
-                AddLineRC("AHSS Air-Reload disabled.");
-            }
-            if (RCSettings.punkWaves > 0)
-            {
-                AddLineRC("Punk override every 5 waves enabled.");
-            }
-            if (RCSettings.endlessMode > 0)
-            {
-                AddLineRC("Endless Respawn is enabled (" + RCSettings.endlessMode.ToString() + " seconds).");
-            }
-            if (RCSettings.globalDisableMinimap > 0)
-            {
-                AddLineRC("Minimap are disabled.");
-            }
-            if (RCSettings.motd != string.Empty)
-            {
-                AddLineRC("MOTD:" + RCSettings.motd);
-            }
-            if (RCSettings.deadlyCannons > 0)
-            {
-                AddLineRC("Cannons will kill humans.");
-            }
-        }
-        else
-        {
-            if (RCSettings.bombMode > 0)
-            {
-                string[] msg = { "Bomb mode is enabled." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.teamMode > 0)
-            {
-                var sort = "Unsorted";
-                if (RCSettings.teamMode == 2)
-                {
-                    sort = "Sorted by size";
-                }
-                else if (RCSettings.teamMode == 3)
-                {
-                    sort = "Sorted by skill";
-                }
-
-                string[] msg = { "Team mode is enabled. ", sort, "." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.pointMode > 0)
-            {
-                string[] msg = { "Points limit is ", $"[{Convert.ToString(RCSettings.pointMode)}]", "." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.disableRock > 0)
-            {
-                string[] msg = { "Punks Rock-Throwing is disabled." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.spawnMode > 0)
-            {
-                string[] msg = { "Custom Spawn Rate is:", $"\n[{RCSettings.nRate.ToString("F2")}% Normal]" + $"\n[{RCSettings.aRate.ToString("F2")}% Abnormal]" + $"\n[{RCSettings.jRate.ToString("F2")}% Jumper]" + $"\n[{RCSettings.cRate.ToString("F2")}% Crawler]" + $"\n[{RCSettings.pRate.ToString("F2")}% Punk]" };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.explodeMode > 0)
-            {
-                string[] msg = { "Explode radius is ", $"[{Convert.ToString(RCSettings.explodeMode)}]", "." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.healthMode > 0)
-            {
-                var mode = "Static ";
-                if (RCSettings.healthMode == 2)
-                {
-                    mode = "Scaled ";
-                }
-
-                string[] msg = { mode + "Health amount is ", $"[{Convert.ToString(RCSettings.healthLower)} - {Convert.ToString(RCSettings.healthUpper)}]", "." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.infectionMode > 0)
-            {
-                string[] msg = { "Infection mode with ", $"[{Convert.ToString(RCSettings.infectionMode)}]", " infected on start." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.damageMode > 0)
-            {
-                string[] msg = { "Minimum Nape Damage is ", $"[{Convert.ToString(RCSettings.damageMode)}]", "." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.moreTitans > 0)
-            {
-                string[] msg = { "Custom Titans Amount is ", $"[{Convert.ToString(RCSettings.moreTitans)}]", "." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.sizeMode > 0)
-            {
-                string[] msg = { "Custom Titans Size is ", $"[{RCSettings.sizeLower.ToString("F2")} - {RCSettings.sizeUpper.ToString("F2")}]", "." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.banEren > 0)
-            {
-                string[] msg = { "Anti-Eren mode is enabled." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.waveModeOn == 1)
-            {
-                string[] msg = { "Custom Titans/Wave amount is ", $"[{Convert.ToString(RCSettings.waveModeNum)}]", "." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.friendlyMode > 0)
-            {
-                string[] msg = { "Friendly mode is enabled." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.pvpMode > 0)
-            {
-                var mode = "";
-                if (RCSettings.pvpMode == 1)
-                {
-                    mode = "Team ";
-                }
-                else if (RCSettings.pvpMode == 2)
-                {
-                    mode = "FFA ";
-                }
-
-                string[] msg = { mode + "PVP mode is enabled." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.maxWave > 0)
-            {
-                string[] msg = { "Custom Maximum Wave is ", $"[{RCSettings.maxWave.ToString()}]", "." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.horseMode > 0)
-            {
-                string[] msg = { "Horses are enabled." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.ahssReload > 0)
-            {
-                string[] msg = { "AHSS Air-Reloading is disabled." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.punkWaves > 0)
-            {
-                string[] msg = { "Punk Waves Override is enabled." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.endlessMode > 0)
-            {
-                string[] msg = { "Endless Respawn is ", $"[{RCSettings.endlessMode.ToString()}]", " seconds." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.globalDisableMinimap > 0)
-            {
-                string[] msg = { "Minimaps are disabled." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.deadlyCannons > 0)
-            {
-                string[] msg = { "Deadly Cannons mode is enabled." };
-                SystemMessageLocal(msg, true);
-            }
-
-            if (RCSettings.motd != string.Empty)
-            {
-                string[] msg = { "MOTD:\n", RCSettings.motd };
-                SystemMessageLocal(msg, false);
-            }
         }
     }
